app/requirements.txt
PyJWT==2.4.0
cryptography==41.0.3
Werkzeug==2.2.2
Flask==2.2.2
Flask-Admin==1.6.0
flask_sqlalchemy==2.5.1
psycopg2-binary==2.9.6
SQLAlchemy==1.4.46